out-source handleInput() & triggerDropDown() functions in Form component to utils file
